refactor: move nullifyTransforms
